Move common database code from mod.rs to new common.rs file
